Fixes, cleanups
