+++ transports/grpc/client/middleware.go
@@ -24,11 +24,14 @@ import (
 
 	grpc_logrus "github.com/grpc-ecosystem/go-grpc-middleware/logging/logrus"
 	"github.com/reeflective/team/client"
-	"github.com/reeflective/team/internal/transport"
 	"github.com/reeflective/team/transports/grpc/common"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 )
+
+// TokenAuth extracts authentication metadata from contexts,
+// specifically the "Authorization": "Bearer" key:value pair.
+type TokenAuth string
 
 // LogMiddlewareOptions is an example list of gRPC options with logging middleware set up.
 // This function uses the core teamclient loggers to log the gRPC stack/requests events.
@@ -69,16 +72,28 @@ func tlsAuthMiddleware(cli *client.Client) ([]grpc.DialOption, error) {
 		return nil, ErrNoTLSCredentials
 	}
 
-	tlsConfig, err := transport.GetTLSConfig(config.CACertificate, config.Certificate, config.PrivateKey)
+	tlsConfig, err := cli.NewTLSConfigFrom(config.CACertificate, config.Certificate, config.PrivateKey)
 	if err != nil {
 		return nil, err
 	}
 
 	transportCreds := credentials.NewTLS(tlsConfig)
-	callCreds := credentials.PerRPCCredentials(transport.TokenAuth(config.Token))
+	callCreds := credentials.PerRPCCredentials(TokenAuth(config.Token))
 
 	return []grpc.DialOption{
 		grpc.WithTransportCredentials(transportCreds),
 		grpc.WithPerRPCCredentials(callCreds),
 	}, nil
+}
+
+// Return value is mapped to request headers.
+func (t TokenAuth) GetRequestMetadata(_ context.Context, _ ...string) (map[string]string, error) {
+	return map[string]string{
+		"Authorization": "Bearer " + string(t),
+	}, nil
+}
+
+// RequireTransportSecurity always return true.
+func (TokenAuth) RequireTransportSecurity() bool {
+	return true
 }
